+++ src/onegov.swissvotes/onegov/swissvotes/forms/search.py
@@ -86,7 +86,7 @@ class SearchForm(Form):
     sort_order = HiddenField()
 
     def populate_policy_area(self):
-        votes = SwissVoteCollection(self.request.session)
+        votes = SwissVoteCollection(self.request.app)
         available = votes.available_descriptors
 
         def serialize(item):